API RESTFUL con CRUD, caché y validaciones

+++ src/main/java/org/example/prueba_spring/crudBasico/Repositorio/IUsuarioRepositorio.java
@@ -2,5 +2,5 @@ package org.example.prueba_spring.crudBasico.Repositorio;
 
 import org.example.prueba_spring.crudBasico.Modelo.Usuario;
 
-public interface IUsuarioRepositorio extends IGenericoRepositorio<Usuario, Long> {
+public interface IUsuarioRepositorio extends IGenericoRepositorio<Usuario, Integer> {
 }
